Removed Grey SVGs. Now use LeftNavSVG to produce grey SVGs. UserBar modified to be stateful component. Sort of updates upvotes live but not really. A few styling changes.
