+++ lambdaface_face/src/Assets/LeftNavCategory/LeftNavSVG.js
@@ -109,7 +109,7 @@ class LeftNavSVG extends React.Component {
               </clipPath>
             </defs>
             <g clipPath="url(#clip_0)">
-              <path fill="rgb(11,11,10)" stroke="none" d="M13.3 10.99C13.5463 10.56 13.7436 10.0975 13.8793 9.61C14.3244 9.495 16 9.015 16 8.1825L16 7.79C16 6.685 14.0063 6.385 13.8818 6.3675C13.7468 5.8775 13.5503 5.415 13.3043 4.985C13.5848 4.6075 14.5499 3.1975 13.8541 2.5025L13.4973 2.125C12.9025 1.53 11.3834 2.46 11.0144 2.7C10.5781 2.4525 10.1093 2.2575 9.61425 2.1225C9.535 1.65 9.18822 0 8.20939 0L7.81761 0C6.97755 0 6.49526 1.7075 6.3875 2.135C5.89946 2.27 5.43717 2.4675 5.00689 2.715C4.66736 2.4625 3.212 1.4575 2.50195 2.1675L2.12492 2.4825C1.50612 3.1025 2.53445 4.735 2.72246 5.02C2.48444 5.445 2.29368 5.9 2.16242 6.38C1.75214 6.445 0 6.7825 0 7.79L0 8.1825C0 9.055 1.83739 9.54 2.17267 9.6225C2.30593 10.0975 2.4977 10.55 2.73621 10.97C2.53845 11.28 1.54987 12.905 2.16342 13.5175L2.5202 13.815C3.32751 14.6225 5.01389 13.2975 5.01389 13.2975L4.93539 13.2125C5.38267 13.4775 5.86696 13.6875 6.37875 13.83C6.4625 14.175 6.94754 16 7.81761 16L8.20939 16C9.35098 16 9.634 13.87 9.634 13.87L9.52924 13.8675C10.0468 13.735 10.5373 13.535 10.9921 13.28C11.3486 13.5175 12.8425 14.4525 13.4388 13.8575L13.8356 13.46C14.6316 12.6625 13.3365 11.035 13.3 10.99ZM8.00212 3.5025C5.51668 3.5025 3.50177 5.5175 3.50177 8.0025C3.50177 10.4875 5.51668 12.5025 8.00212 12.5025C10.4876 12.5025 12.5025 10.4875 12.5025 8.0025C12.5025 5.5175 10.4876 3.5025 8.00212 3.5025ZM5.00189 8.0025C5.00189 9.66 6.34525 11.0025 8.00212 11.0025C9.65925 11.0025 11.0024 9.66 11.0024 8.0025C11.0024 6.345 9.65925 5.0025 8.00212 5.0025C6.34525 5.0025 5.00189 6.345 5.00189 8.0025ZM6.50201 8.0025C6.50201 7.175 7.17381 6.5025 8.00212 6.5025C8.83069 6.5025 9.50224 7.175 9.50224 8.0025C9.50224 8.83 8.83069 9.5025 8.00212 9.5025C7.17381 9.5025 6.50201 8.83 6.50201 8.0025Z" fillRule="evenodd" />
+              <path fill={this.props.fill} stroke="none" d="M13.3 10.99C13.5463 10.56 13.7436 10.0975 13.8793 9.61C14.3244 9.495 16 9.015 16 8.1825L16 7.79C16 6.685 14.0063 6.385 13.8818 6.3675C13.7468 5.8775 13.5503 5.415 13.3043 4.985C13.5848 4.6075 14.5499 3.1975 13.8541 2.5025L13.4973 2.125C12.9025 1.53 11.3834 2.46 11.0144 2.7C10.5781 2.4525 10.1093 2.2575 9.61425 2.1225C9.535 1.65 9.18822 0 8.20939 0L7.81761 0C6.97755 0 6.49526 1.7075 6.3875 2.135C5.89946 2.27 5.43717 2.4675 5.00689 2.715C4.66736 2.4625 3.212 1.4575 2.50195 2.1675L2.12492 2.4825C1.50612 3.1025 2.53445 4.735 2.72246 5.02C2.48444 5.445 2.29368 5.9 2.16242 6.38C1.75214 6.445 0 6.7825 0 7.79L0 8.1825C0 9.055 1.83739 9.54 2.17267 9.6225C2.30593 10.0975 2.4977 10.55 2.73621 10.97C2.53845 11.28 1.54987 12.905 2.16342 13.5175L2.5202 13.815C3.32751 14.6225 5.01389 13.2975 5.01389 13.2975L4.93539 13.2125C5.38267 13.4775 5.86696 13.6875 6.37875 13.83C6.4625 14.175 6.94754 16 7.81761 16L8.20939 16C9.35098 16 9.634 13.87 9.634 13.87L9.52924 13.8675C10.0468 13.735 10.5373 13.535 10.9921 13.28C11.3486 13.5175 12.8425 14.4525 13.4388 13.8575L13.8356 13.46C14.6316 12.6625 13.3365 11.035 13.3 10.99ZM8.00212 3.5025C5.51668 3.5025 3.50177 5.5175 3.50177 8.0025C3.50177 10.4875 5.51668 12.5025 8.00212 12.5025C10.4876 12.5025 12.5025 10.4875 12.5025 8.0025C12.5025 5.5175 10.4876 3.5025 8.00212 3.5025ZM5.00189 8.0025C5.00189 9.66 6.34525 11.0025 8.00212 11.0025C9.65925 11.0025 11.0024 9.66 11.0024 8.0025C11.0024 6.345 9.65925 5.0025 8.00212 5.0025C6.34525 5.0025 5.00189 6.345 5.00189 8.0025ZM6.50201 8.0025C6.50201 7.175 7.17381 6.5025 8.00212 6.5025C8.83069 6.5025 9.50224 7.175 9.50224 8.0025C9.50224 8.83 8.83069 9.5025 8.00212 9.5025C7.17381 9.5025 6.50201 8.83 6.50201 8.0025Z" fillRule="evenodd" />
             </g>
           </svg>
         )
