+++ lambdaface_face/src/Styles/less/index.less
@@ -14,6 +14,5 @@
 @import "landingPage";
 @import "writeComment";
 @import "writeReply";
-@import "leftNav";
 @import "comment";
 @import "markdown";
