+++ lambdaface_face/src/Styles/less/userBar.less
@@ -31,6 +31,9 @@
       fill: grey;
     }
   }
+  &__category-name {
+    margin-left: 3px;
+  }
   &__post-details {
     display: flex;
     width: 55%;
